Remove obsolete config option from language strings Added missing language string for config option Fixed wrong language string for config options Move legacy config option mapper out of config model and add the enabled settings to it Fixed wrong description
