+++ HydrateOrDiedrate/src/Config/ModConfig.cs
@@ -76,37 +76,4 @@ public class ModConfig
 
     [JsonExtensionData]
     public Dictionary<string, JToken> LegacyData { get; set; }
-
-    public bool MapLegacyData()
-    {
-        if(LegacyData is null) return false;
-        
-        var legacyData = new Dictionary<string, JToken>(LegacyData, StringComparer.OrdinalIgnoreCase);
-        var subConfigs = typeof(ModConfig)
-            .GetProperties()
-            .Where(prop => prop.PropertyType.IsClass)
-            .Select(prop => prop.GetValue(this));
-
-        foreach(var subConfig in subConfigs)
-        {
-            foreach(var property in subConfig.GetType().GetProperties())
-            {
-                if(legacyData.TryGetValue(property.Name, out var token))
-                {
-                    try
-                    {
-                        property.SetValue(subConfig, token.ToObject(property.PropertyType));
-                    }
-                    catch
-                    {
-                        //Ignore unknown legacy data
-                    }
-                }
-            }
-        }
-
-        LegacyData = null;
-        return true;
-    }
-
 }
